finish work

+++ src/main/java/teampcc/backendservice/repository/BossRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import teampcc.backendservice.entity.BossEntity;
 
 
-public interface BossRepository extends JpaRepository<BossEntity,Integer> {
+public interface BossRepository extends JpaRepository<BossEntity,Long> {
 
 }
